pyproject.toml
[project]
name = "arxiv-to-mp3"
version = "0.1.0"
description = "Convert arXiv papers into podcast audio"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "albumentations==1.0.0",
  "google-cloud-storage>=3.1.0",
  "nougat-ocr>=0.1.17",
  "openai>=1.75.0",
  "pulumi>=3.162.0",
  "pulumi-gcp>=8.26.0",
  "pypdf2>=3.0.1",
  "transformers==4.38.2",
]
